refactor(access-service): extract admin/ops read-surface into access-admin-ops-surface.ts
God-file decomposition (#1526 playbook). Moves the operator-facing
read/maintenance surfaces (771 LOC, 8 methods): reviewQueue,
governanceRun, conversationIndexUpdate, profilingReport, graphSnapshot,
capsuleList, dreamsRun, adminPromoteMemory into AccessAdminOpsSurface
(3 deps via selfDeps live wiring). Governance projection helpers and
profiling types exported from access-service for the surface module.

Stacked on refactor/access-service-recall-surface.

access-service.ts: 7,849 -> 7,192 LOC.
